requirements.txt
appnope==0.1.0
backcall==0.1.0
bcrypt==3.1.6
cffi==1.12.2
decorator==4.4.0
Django==1.10
ipython==7.4.0
ipython-genutils==0.2.0
jedi==0.13.3
parso==0.4.0
pexpect==4.7.0
pickleshare==0.7.5
prompt-toolkit==2.0.9
ptyprocess==0.6.0
pycparser==2.19
Pygments==2.3.1
python-dateutil==2.8.0
pytz==2018.9
six==1.12.0
traitlets==4.3.2
wcwidth==0.1.7
